@@ -1,22 +1,22 @@
 name: auto_route_example
 publish_to: 'none'
-version: 1.0.0-beta
+version: 1.0.0-1
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   get_it: ^6.0.0
-  auto_route: ^1.0.0-beta
+  auto_route: ^1.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  auto_route_generator: ^1.0.0-beta
+  auto_route_generator: ^1.0.0
   build_runner: ^1.0.0
 
 flutter:
