@@ -6,7 +6,7 @@ import 'app_router.gr.dart';
 class HomeRouterPage extends StatefulWidget {
   HomeRouterPage({this.pageId});
 
-  final String pageId;
+  final String? pageId;
 
   @override
   _HomeRouterPageState createState() => _HomeRouterPageState();
@@ -15,20 +15,18 @@ class HomeRouterPage extends StatefulWidget {
 class _HomeRouterPageState extends State<HomeRouterPage> {
   @override
   Widget build(BuildContext context) => AutoRouter.declarative(
-    onGenerateRoutes: (context, routes) {
-      return [
-        HomeRoute(),
-        if (widget.pageId == 'a') HomeARoute(),
-        if (widget.pageId == 'b') HomeBRoute(),
-      ];
-    },
-  );
+        onGenerateRoutes: (context, routes) => [
+          HomeRoute(),
+          if (widget.pageId == 'a') HomeARoute(),
+          if (widget.pageId == 'b') HomeBRoute(),
+        ],
+      );
 }
 
 class PublicRouterPage extends StatefulWidget {
   PublicRouterPage({this.pageId});
 
-  final String pageId;
+  final String? pageId;
 
   @override
   _PublicRouterPageState createState() => _PublicRouterPageState();
@@ -37,20 +35,18 @@ class PublicRouterPage extends StatefulWidget {
 class _PublicRouterPageState extends State<PublicRouterPage> {
   @override
   Widget build(BuildContext context) => AutoRouter.declarative(
-    onGenerateRoutes: (context, routes) {
-      return [
-        PublicRoute(),
-        if (widget.pageId == 'a') PublicARoute(),
-        if (widget.pageId == 'b') PublicBRoute(),
-      ];
-    },
-  );
+        onGenerateRoutes: (context, routes) => [
+          PublicRoute(),
+          if (widget.pageId == 'a') PublicARoute(),
+          if (widget.pageId == 'b') PublicBRoute(),
+        ],
+      );
 }
 
 class ProtectedRouterPage extends StatefulWidget {
   ProtectedRouterPage({this.pageId});
 
-  final String pageId;
+  final String? pageId;
 
   @override
   _ProtectedRouterPageState createState() => _ProtectedRouterPageState();
@@ -59,12 +55,10 @@ class ProtectedRouterPage extends StatefulWidget {
 class _ProtectedRouterPageState extends State<ProtectedRouterPage> {
   @override
   Widget build(BuildContext context) => AutoRouter.declarative(
-        onGenerateRoutes: (context, routes) {
-          return [
-            ProtectedRoute(),
-            if (widget.pageId == 'a') ProtectedARoute(),
-            if (widget.pageId == 'b') ProtectedBRoute(),
-          ];
-        },
+        onGenerateRoutes: (context, routes) => [
+          ProtectedRoute(),
+          if (widget.pageId == 'a') ProtectedARoute(),
+          if (widget.pageId == 'b') ProtectedBRoute(),
+        ],
       );
 }
